Added access point index to the base station structure.

+++ aether/wifi/esp_wifi_driver.h
@@ -37,7 +37,7 @@ class EspWifiDriver final : public WifiDriver {
 
   void Connect(WiFiInit& wifi_init, WiFiBaseStation& base_station_) override;
   WifiCreds connected_to() const override;
-  size_t ap_cnt() const override;
+  uint16_t ap_cnt() const override;
 
  private:
   void Disconnect();
@@ -48,7 +48,7 @@ class EspWifiDriver final : public WifiDriver {
 
   static int initialized_;
   static void* espt_init_sta_;
-  size_t ap_cnt_;
+  uint16_t ap_cnt_;
 
   std::optional<WifiCreds> connected_to_;
   std::unique_ptr<struct ConnectionState> connection_state_;
